feat: prepare usage foundation

+++ src/main/java/com/watchmyai/quota/PlanLimits.java
@@ -1,11 +1,13 @@
 package com.watchmyai.quota;
 
+import java.math.BigDecimal;
+
 public record PlanLimits(
         PlanType planType,
         int lifetimeRequestLimit,
         int monthlyRequestLimit,
         int monthlyPremiumRequestLimit,
         int maxOutputTokens,
-        double monthlyCostCapEur
+        BigDecimal monthlyCostCapEur
 ) {
 }
